tests/submission/MOSQUITO/event-source/user-agent-announce-connection.html

<!DOCTYPE html>
<html>
<head>
	<title>When a user agent is to announce the connection, the user agent must queue a task which, if the readyState attribute is set to a value other than CLOSED, sets the readyState attribute to OPEN and fires a simple event named open at the EventSource object:</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>When a user agent is to announce the connection, the user agent must queue a task which, if the readyState attribute is set to a value other than CLOSED, sets the readyState attribute to OPEN and fires a simple event named open at the EventSource object:</h1>
	<div id="log"></div>
  <script>
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0051');
    
    s.onopen = function(e){ 
			/* Connection opened. */
			ttop.step(function(){assert_true(s.readyState == 1)});
			ttop.done();
    };
    test(function(){assert_true(s.readyState != 2)},"The readyState attribute is set to a value other than 'CLOSED' (2)");
    ttop = async_test("The readyState attribute is set to 'OPEN' (1)");
    test(function(){assert_true(typeof(s.onopen) == "function")},"Fires a simple event named open at the EventSource object.");
    
  </script>
  </div>
</body>
</html>
